requirements.txt
requests==2.32.2
Jinja2==3.1.4
Pillow==10.3.0